+++ request/Post/Chunked.hpp
@@ -26,4 +26,4 @@ public:
     int handleChunked();
     void initializeMimeTypes();
 };
-int pasteInFile(std::string name, std::string &data);
+size_t pasteInFile(std::string name, std::string &data);
